feat: add the delete character feature

+++ backend-challenge/src/main/java/com/macedo/backend_challenge/application/services/CharacterService.java
@@ -80,6 +80,12 @@ public class CharacterService {
     }
 
     public void remove(Integer id) {
+        var character = characterRepository.getCharacterById(id);
+
+        if (character == null) {
+            throw new CharacterNotFoundException();
+        }
+
         characterRepository.deleteById(id);
     }
